[Teseo] Explicitly set the thread affinity
It turns out that the env. vars for the OpenMP thread affinity are
completely ignored by Clang and have a global effect, that is, they
affect all threads, in GCC. With this patch, avoid using the env.
vars for OpenMP to set the thread affinity, but set it manually
through proper pthread function calls upon necessity.

+++ tests/test_updates_directed.cpp
@@ -226,7 +226,7 @@ static void parallel(shared_ptr<UpdateInterface> interface, uint64_t num_vertice
         threads.clear();
         start = 0;
         for(int thread_id = 0; thread_id < num_threads; thread_id ++){
-            uint64_t length = vertices_per_thread + (thread_id < odd_threads);
+            uint64_t length = vertices_per_thread + ((uint64_t) thread_id < odd_threads);
             threads.emplace_back(routine_remove_vertices, thread_id, start, length);
             start += length;
         }
